+++ src/main/java/codesquad/domain/QuestionPageable.java
@@ -1,0 +1,46 @@
+package codesquad.domain;
+
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+
+public class QuestionPageable implements Pageable {
+    @Override
+    public int getPageNumber() {
+        return 0;
+    }
+
+    @Override
+    public int getPageSize() {
+        return 15;
+    }
+
+    @Override
+    public int getOffset() {
+        return 0;
+    }
+
+    @Override
+    public Sort getSort() {
+        return null;
+    }
+
+    @Override
+    public Pageable next() {
+        return null;
+    }
+
+    @Override
+    public Pageable previousOrFirst() {
+        return null;
+    }
+
+    @Override
+    public Pageable first() {
+        return null;
+    }
+
+    @Override
+    public boolean hasPrevious() {
+        return false;
+    }
+}
