+++ src/main/java/codesquad/web/HomeController.java
@@ -1,10 +1,12 @@
 package codesquad.web;
 
+import codesquad.domain.QuestionPageable;
 import codesquad.domain.QuestionRepository;
 import com.github.jknack.handlebars.Handlebars;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,8 +20,9 @@ public class HomeController {
     private QuestionRepository questionRepository;
 
     @GetMapping({"/", "index"})
-    public String home(Model model) {
-        model.addAttribute("posts", questionRepository.findAll(new Sort(Sort.Direction.DESC, "id")));
+    public String home(Model model, QuestionPageable questionPageable) {
+//        model.addAttribute("posts", questionRepository.findAll(new Sort(Sort.Direction.DESC, "id")));
+        model.addAttribute("posts", questionRepository.findAll(questionPageable));
         log.debug("/index/success");
         return "index";
     }
